🤖 refactor: volunteerEvent 인터페이스 types 폴더로 이동
