+++ src/app/page.tsx
@@ -1,7 +1,6 @@
 'use client';
 import HomePage from '@/components/Home/HomePage';
-import ScheduleTab from '../components/volunteer-schedule/ScheduleTab/ScheduleTab';
 
 export default function Home() {
-  return <ScheduleTab />;
+  return <HomePage />;
 }
