+++ src/components/volunteer-schedule/mock.ts
@@ -1,4 +1,4 @@
-import { VolunteerEvent } from './VolunteerEventCard/VolunteerEventCard';
+import { VolunteerEvent } from '../../types/volunteerEvent';
 
 export const volunteerEventsMock: VolunteerEvent[] = [
   {
@@ -7,7 +7,7 @@ export const volunteerEventsMock: VolunteerEvent[] = [
     volunteerEventId: 11,
     title: '한강 인근 댕댕이 산책 봉사자 모집합니다.',
     recruitNum: 5,
-    participantNum: 5,
+    joinNum: 5,
     waitingNum: 2,
     startAt: '2023-07-01T12:30:55.162Z',
     endAt: '2023-07-01T13:30:55.162Z',
@@ -19,7 +19,7 @@ export const volunteerEventsMock: VolunteerEvent[] = [
     volunteerEventId: 12,
     title: '태평역 인근 산책 봉사자 모집합니다.',
     recruitNum: 5,
-    participantNum: 5,
+    joinNum: 5,
     waitingNum: 5,
     startAt: '2023-07-01T19:00:55.162Z',
     endAt: '2023-07-01T20:00:55.162Z',
@@ -30,7 +30,7 @@ export const volunteerEventsMock: VolunteerEvent[] = [
     category: '견사 청소',
     volunteerEventId: 13,
     title: '노원역 인근 견사 청소 봉사자 모집합니다.',
-    participantNum: 4,
+    joinNum: 4,
     recruitNum: 5,
     waitingNum: 0,
     startAt: '2023-07-09T12:30:55.162Z',
@@ -43,7 +43,7 @@ export const volunteerEventsMock: VolunteerEvent[] = [
     volunteerEventId: 14,
     title: '홍보물 제작 봉사자 모집합니다.',
     recruitNum: 6,
-    participantNum: 6,
+    joinNum: 6,
     waitingNum: 3,
     startAt: '2023-07-09T10:00:55.162Z',
     endAt: '2023-07-09T12:00:55.162Z',
@@ -55,7 +55,7 @@ export const volunteerEventsMock: VolunteerEvent[] = [
     volunteerEventId: 15,
     title: '홍보물 제작 봉사자 모집합니다.',
     recruitNum: 6,
-    participantNum: 6,
+    joinNum: 6,
     waitingNum: 3,
     startAt: '2023-07-20T12:00:55.162Z',
     endAt: '2023-07-20T13:00:55.162Z',
@@ -67,7 +67,7 @@ export const volunteerEventsMock: VolunteerEvent[] = [
     volunteerEventId: 16,
     title: '홍보물 제작 봉사자 모집합니다.',
     recruitNum: 6,
-    participantNum: 6,
+    joinNum: 6,
     waitingNum: 3,
     startAt: '2023-07-29T12:00:55.162Z',
     endAt: '2023-07-29T16:00:55.162Z',
